refactor: renaming the validations to adapt to the standard
