+++ src/main/factories/validations/tasks/index.ts
@@ -1,3 +1,3 @@
-export * from './tagId-validation-factory'
+export * from './load-current-date-tasks-validation-factory'
+export * from './load-current-week-tasks-validation-factory'
 export * from './load-current-month-tasks-validation-factory'
-export * from './load-tasks-by-current-week-validation-factory'
