+++ src/main/factories/validations/tasks/load-current-week-tasks-validation-factory.ts
@@ -2,7 +2,7 @@ import { ValidationComposite, TagIdValidation } from '@/application/validation'
 import { UuidValidatorAdapter } from '@/infra/validators'
 import { type Validation } from '@/application/protocols'
 
-export const makeLoadTasksByCurrentWeekValidation = (): ValidationComposite => {
+export const makeLoadCurrentWeekTasksValidation = (): ValidationComposite => {
   const validations: Validation[] = []
 
   validations.push(new TagIdValidation('tagId', new UuidValidatorAdapter()))
